Updated sample codes
