+++ GarlicWebviewSample/Assets/GarlicWebview/scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 40d1e48074d6f1841a44dcde3f0b105b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
