+++ GarlicWebviewSample/Assets/GarlicWebview/scripts/GarlicAndroidUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 518a80960e6e90e42b666f106c8bf9e7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
