+++ GarlicWebviewSample/Assets/GarlicWebview/scripts/impl.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 51b905eb371eaaf40bcae65d6f7881e7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
