+++ GarlicWebviewSample/Assets/GarlicWebview/scripts/impl/GarlicWebviewAndroidImpl.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 48dd19e72b0fbf948a9115931d066f1a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
