+++ GarlicWebviewSample/Assets/GarlicWebview/scripts/impl/GarlicWebviewiOSImpl.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d9b27481c17cbe34b84fba45e3902302
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
