+++ GarlicWebviewSample/Assets/Plugins/Android/mainTemplate.gradle
@@ -1,0 +1,83 @@
+// GENERATED BY UNITY. REMOVE THIS COMMENT TO PREVENT OVERWRITING WHEN EXPORTING AGAIN
+buildscript {
+	repositories {
+		google()
+		jcenter()
+	}
+
+	dependencies {
+		classpath 'com.android.tools.build:gradle:3.2.0'
+	}
+}
+
+allprojects {
+   repositories {
+      google()
+      jcenter()
+      flatDir {
+        dirs 'libs'
+      }
+   }
+}
+
+apply plugin: 'com.android.application'
+
+dependencies {
+	implementation fileTree(dir: 'libs', include: ['*.jar'])
+	implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+**DEPS**}
+
+android {
+	compileSdkVersion **APIVERSION**
+	buildToolsVersion '**BUILDTOOLS**'
+
+	defaultConfig {
+		minSdkVersion **MINSDKVERSION**
+		targetSdkVersion **TARGETSDKVERSION**
+		applicationId '**APPLICATIONID**'
+		ndk {
+            abiFilters **ABIFILTERS**
+        }
+		versionCode	**VERSIONCODE**
+		versionName	'**VERSIONNAME**'
+	}
+
+	lintOptions {
+		abortOnError false
+	}
+
+	aaptOptions {
+		noCompress '.unity3d', '.ress', '.resource', '.obb'**STREAMING_ASSETS**
+	}
+
+**SIGN**
+	buildTypes {
+  		debug {
+ 			minifyEnabled **MINIFY_DEBUG**
+ 			useProguard **PROGUARD_DEBUG**
+ 			proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-unity.txt'**USER_PROGUARD**
+  			jniDebuggable true
+  		}
+  		release {
+ 			minifyEnabled **MINIFY_RELEASE**
+ 			useProguard **PROGUARD_RELEASE**
+  			proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-unity.txt'**USER_PROGUARD**
+  			**SIGNCONFIG**
+  		}
+	}
+**PACKAGING_OPTIONS**
+    bundle {
+        language {
+            enableSplit = false
+        }
+        density {
+            enableSplit = false
+        }
+        abi {
+            enableSplit = true
+        }
+    }
+}
+
+
+**SOURCE_BUILD_SETUP**
