+++ GarlicWebviewSample/Assets/Plugins/Android/mainTemplate.gradle.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c9c55b612af20b54eb7a49353b88c6c3
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
